refactor: result 패키지 분리
